+++ packages/outbox/src/test/utils/helpers/event/event-queue-pusher-mock.helper.ts
@@ -1,6 +1,6 @@
 import type { jest } from '@jest/globals';
 import { createMock } from '@volontariapp/database/testing';
-import type { EventQueuePusher } from '../../../pushers/event-queue.pusher.js';
+import type { EventQueuePusher } from '../../../../pushers/event-queue.pusher.js';
 
 export type EventQueuePusherMock = jest.Mocked<EventQueuePusher>;