+++ packages/outbox/src/test/utils/helpers/event/event-queue-repository-mock.helper.ts
@@ -3,11 +3,11 @@ import { EventQueueModel } from '@volontariapp/database';
 import {
   makeOutboxRepositoryMock,
   type OutboxRepositoryMock,
-} from './outbox-repository-mock.helper.js';
+} from '../shared/outbox-repository-mock.helper.js';
 import {
   makeOutboxConsumerRepositoryMock,
   type OutboxConsumerRepositoryMock,
-} from './outbox-consumer-repository-mock.helper.js';
+} from '../shared/outbox-consumer-repository-mock.helper.js';
 
 export type EventQueueRepositoryMock = OutboxRepositoryMock<EventQueueEntity, EventQueueModel>;