+++ packages/outbox/src/test/writers/event/event-queue.writer.unit.spec.ts
@@ -6,12 +6,12 @@ import {
   type EventQueueModel,
 } from '@volontariapp/database';
 import { EventQueueWriter } from '../../../writers/event-queue.writer.js';
-import { makeEventQueueEvent } from '../../utils/helpers/event-queue-event.helper.js';
-import { makeLoggerMock } from '../../utils/helpers/logger-mock.helper.js';
+import { makeEventQueueEvent } from '../../utils/helpers/event/event-queue-event.helper.js';
+import { makeLoggerMock } from '../../utils/helpers/shared/logger-mock.helper.js';
 import {
   makeEventQueueRepositoryMock,
   type EventQueueRepositoryMock,
-} from '../../utils/helpers/event-queue-repository-mock.helper.js';
+} from '../../utils/helpers/event/event-queue-repository-mock.helper.js';
 
 describe('EventQueueWriter (Unit)', () => {
   let writer: EventQueueWriter;