+++ packages/shared/src/enums/index.ts
@@ -1,2 +1,3 @@
 export * from './tags-enum.js';
 export * from './user-roles-enum.js';
+export * from './service-type.enum.js';