+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.4.0
+
+### Minor Changes
+
+- Event queue pusher added
+
 ## 0.3.0
 
 ### Minor Changes